# Conflicts:
#	src/main/java/pl/weather/api/CloudApi.java
#	src/main/java/pl/weather/api/UrlProvider.java
#	src/main/java/pl/weather/api/WeatherRequestBuilder.java
#	src/main/java/pl/weather/models/Clouds.java
#	src/main/java/pl/weather/models/Coord.java
#	src/main/java/pl/weather/models/Details.java
#	src/main/java/pl/weather/models/SimpleDataWeatherConverter.java
#	src/main/java/pl/weather/models/Sys.java
#	src/main/java/pl/weather/models/Weather.java
#	src/main/java/pl/weather/models/WeatherForCity.java
#	src/main/java/pl/weather/models/Wind.java
#	src/main/java/pl/weather/storage/DataFileManager.java
#	src/main/java/pl/weather/storage/FileStorage.java
#	src/main/java/pl/weather/tools/intervalchecker/WeatherIntervalCheckTask.java

+++ src/main/java/pl/weather/models/Clouds.java
@@ -4,17 +4,15 @@ import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
 public class Clouds {
-
     @SerializedName("all")
     @Expose
-    private int all;
+    private double all;
 
-    public int getAll() {
+    public double getAll() {
         return all;
     }
 
-    public void setAll(int all) {
+    public void setAll(double all) {
         this.all = all;
     }
-
 }
