+++ src/main/java/pl/weather/models/Wind.java
@@ -1,17 +1,16 @@
-
 package pl.weather.models;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
 public class Wind {
-
     @SerializedName("speed")
     @Expose
     private double speed;
+
     @SerializedName("deg")
     @Expose
-    private int deg;
+    private double deg;
 
     public double getSpeed() {
         return speed;
@@ -21,12 +20,11 @@ public class Wind {
         this.speed = speed;
     }
 
-    public int getDeg() {
+    public double getDeg() {
         return deg;
     }
 
-    public void setDeg(int deg) {
+    public void setDeg(double deg) {
         this.deg = deg;
     }
-
 }
